Cleaned up control group files
Refactor
