+++ src/plugins/controls/common/control_group/control_group_constants.ts
@@ -8,6 +8,6 @@
 
 import { ControlStyle, ControlWidth } from '../types';
 
-export const DEFAULT_CONTROL_WIDTH: ControlWidth = 'medium';
 export const DEFAULT_CONTROL_GROW: boolean = true;
 export const DEFAULT_CONTROL_STYLE: ControlStyle = 'oneLine';
+export const DEFAULT_CONTROL_WIDTH: ControlWidth = 'medium';
