+++ src/plugins/controls/common/control_group/control_group_persistable_state.ts
@@ -6,15 +6,16 @@
  * Side Public License, v 1.
  */
 
+import { SavedObjectReference } from '@kbn/core/types';
 import {
   EmbeddableInput,
   EmbeddablePersistableStateService,
   EmbeddableStateWithType,
 } from '@kbn/embeddable-plugin/common/types';
-import { SavedObjectReference } from '@kbn/core/types';
 import { MigrateFunctionsObject } from '@kbn/kibana-utils-plugin/common';
-import { ControlGroupInput, ControlPanelState } from './types';
+
 import { makeControlOrdersZeroBased } from './control_group_migrations';
+import { ControlGroupInput, ControlPanelState } from './types';
 
 type ControlGroupInputWithType = Partial<ControlGroupInput> & { type: string };
