+++ src/plugins/controls/server/control_group/control_group_container_factory.ts
@@ -8,6 +8,7 @@
 
 import { EmbeddablePersistableStateService } from '@kbn/embeddable-plugin/common';
 import { EmbeddableRegistryDefinition } from '@kbn/embeddable-plugin/server';
+
 import { CONTROL_GROUP_TYPE } from '../../common';
 import {
   createControlGroupExtract,
